+++ src/entities/message/Audio.php
@@ -25,6 +25,7 @@ class Audio extends BaseObject
     public int $duration;
     public ?string $performer = null;
     public ?string $title = null;
+    public ?string $file_name = null;
     public ?string $mime_type = null;
     public ?int $file_size = null;
     public ?PhotoSize $thumb = null;