+++ src/entities/webhook/WebhookInfo.php
@@ -19,6 +19,7 @@ class WebhookInfo extends BaseObject
     public string $url;
     public bool $has_custom_certificate;
     public int $pending_update_count;
+    public string $ip_address;
     public int $last_error_date;
     public string $last_error_message;
     public int $max_connections;